PROSIT 5 || Q5 done

+++ src/tn/tuniprod/gestionmagasin/Magasin.java
@@ -165,8 +165,8 @@ public class Magasin {
     }
 
     //PROSIT 5
-    public float calculStock(){
-        float quantiteTotale=0;
+    public int calculStock(){
+        int quantiteTotale=0;
         for (int i = 0; i < this.capacite; i++) {
             if (this.produits[i] instanceof ProduitFruit){
                quantiteTotale +=1;
